GAL-18 Rename Pages folder into pages
