+++ packages/client/src/pages/Profile/components/Avatar/index.module.scss
