+++ packages/client/src/pages/Topic/components/SmileMenu/index.module.scss
